requirements.txt
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
# redis>=5.0.0  # Removed - not needed for current architecture
celery>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.3.0
anthropic>=0.7.0 